+++ code/system.ui/react.common/src/ui/Button/Root.tsx
@@ -1,0 +1,11 @@
+import { DEFAULTS, FC, type t } from './common';
+import { View } from './ui.Button';
+
+type Fields = {
+  DEFAULTS: typeof DEFAULTS;
+};
+export const Button = FC.decorate<t.ButtonProps, Fields>(
+  View,
+  { DEFAULTS },
+  { displayName: 'Button' },
+);
