+++ code/system.ui/react.common/src/ui/Button/index.ts
@@ -1,1 +1,1 @@
-export * from './Button';
+export * from './Root';
